Hay que borrar ficheros de errores

+++ programar/Cursus/Cirulo0/libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: abastard <abastard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/16 18:40:34 by abastard          #+#    #+#             */
-/*   Updated: 2024/02/24 12:54:04 by abastard         ###   ########.fr       */
+/*   Updated: 2024/02/24 17:01:18 by abastard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strstr(char *pajar, char *aguja, __SIZE_TYPE__ n)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
